# api/openapi.yaml
openapi: 3.0.0
info:
  title: Hyprtime Daemon API
  description: |
    REST API for the Hyprtime daemon that tracks Hyprland window usage.
    The API is exposed over a Unix socket for fast, secure local IPC.
  version: 1.0.0
  contact:
    name: Hyprtime
    url: https://github.com/ayanrajpoot10/hyprtime

servers:
  - url: http://unix
    description: Unix socket at $XDG_RUNTIME_DIR/hyprtime/daemon.sock

paths:
  /api/v1/health:
    get:
      summary: Health check
      description: Returns the health status of the daemon API
      operationId: getHealth
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  time:
                    type: string
                    format: date-time
                    example: "2026-03-21T12:00:00Z"

  /api/v1/stats/today:
    get:
      summary: Get today's stats
      description: Returns screen time statistics for the current day
      operationId: getTodayStats
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyData'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/stats/daily/{date}:
    get:
      summary: Get daily stats
      description: Returns screen time statistics for a specific date
      operationId: getDailyStats
      parameters:
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
            example: "2026-03-21"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyData'
        '400':
          description: Invalid date format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    DailyData:
      type: object
      description: Daily screen time statistics
      properties:
        date:
          type: string
          format: date
          description: Date in YYYY-MM-DD format
          example: "2026-03-21"
        total_time:
          type: integer
          description: Total screen time in seconds
          example: 3600
        total_time_formatted:
          type: string
          description: Human-readable total time
          example: "1h 0m 0s"
        apps:
          type: array
          description: List of applications used
          items:
            $ref: '#/components/schemas/AppData'

    AppData:
      type: object
      description: Application usage statistics
      properties:
        class:
          type: string
          description: Application class name
          example: "kitty"
        total_time:
          type: integer
          description: Time spent in seconds
          example: 1800
        total_time_formatted:
          type: string
          description: Human-readable time
          example: "30m 0s"
        open_count:
          type: integer
          description: Number of times app was opened
          example: 5
        last_seen:
          type: string
          format: date
          description: Last date app was used
          example: "2026-03-21"
        percentage:
          type: number
          format: float
          description: Percentage of total screen time
          example: 50.0

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Database error: connection failed"